requirements.txt
streamlit
spacy
nltk
langdetect
pypdf
python-docx
reportlab
